+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Skip `TrackAppExt::track_mutate_messages` if it has already been called.
+
 ## [0.37.0] - 2025-12-14
 
 ### Fixed